--- TO_DO.txt
@@ -1,4 +1,0 @@
-JSONs Ingles Español
-
-Fetch de los JSONs
-
