--- public/data/data-en.json
@@ -1,93 +1,0 @@
-{
-        "aboutMe": "About me",
-        "studies": "Studies",
-        "skills": "Skills",
-        "work": "Work",
-        "projects": "Projects",
-       "aboutMeTitle": "Hi I’m Alex!",
-       "aboutMeSubtitle":"Web developer and Stage designer",
-       "aboutMePhrase":"As a recent graduate in Web Development, I'm passionate about creating impactful, user-centered websites and applications. I have developed a solid foundation in back-end and front-end technologies.",
-        "studiesTitle": "Studies",
-        "studies1": {
-                "title": "Higher Technician in Digital design and Web Development.",
-                "description": "I've learn the use of basic developing lenguages and developed some projects in this area. Mostly worked with Java adn Spring Boot, with MySQL DB and HTML, CSS and JS. Also worked with Git and Github and deploying in Azure.",
-                "date": "Santa Joaquina de Vedruna (2023-2025)"
-                },
-        "studies2": {
-                "title": "University degree in Stage Design and Scenography Technology.",
-                "description": "Spend 4 years designing and working in groups for different theater proyects, with a strong focus on design, production and team management",
-                "date": "Escuela Superior de Arte Dramático  de Sevilla (2017-2022)"
-                },
-        "studies3": {
-                        "title": "French language medium level",
-                        "description": "Studied French language at the Instituto de Idiomas de la Universidad de Sevilla for several months and reached a B2 level.",
-                        "date": "Instituto de Idiomas de la Universidad de Sevilla (2021)"
-                        },
-        
-
-        "skillsTitle": "Skills",
-        "hardskills": "Hard Skills",
-        "softskills": "Soft Skills",
-        "workTitle": "Work",
-        "work1": {
-                "title": "NTTDATA Internship",
-                "description": "Spent 3 months as a NTTDATA Dual Intern with the role of Junior Web Developer.",
-                "date": "2023",
-                "link": "https://es.nttdata.com"
-                },
-        "work2": {
-                "title": "Veganitessen Crew Leader",
-                "description": "Worked as a Veganitessen Crew Leader at the university.",
-                "date": "2020-2021"
-                },
-        "referencesTitle": "References",
-        "reference1": {
-                "img": "/references/reference1.png",
-                "title": "Jose Martínez",
-                "description": "Profesor de diseño de la escena",
-                "contact": "Contacto: jose@telefónica.es"
-                },
-        "reference2": {
-                "img": "/references/reference2.png",
-                "title": "Anabel Reyes",
-                "description": "Coordinadora del restaurante Veganitessen",
-                "contact": "Contacto: anabelReyes@gmail.com"
-                },
-        "reference3": {
-                "img": "/references/reference3.png",
-                "title": "José Navarro",
-                "description": "Jefe de equipo de montaje Wallahye",
-                "contact": "Contacto: wallahyeProd@hotmail.es"
-                },
-        "reference4": {
-                "img": "/references/reference4.png",
-                "title": "Joaquín Borrego",
-                "description": "Tutor del grado DAW en Vedruna",
-                "contact": "Contacto: joaquin@telefónica.es"
-                },
-        "projectTitle": "Projects",
-        "buttonNext": "Next",
-        "buttonPrev": "Previous",
-        "buttonSearch": "Search",
-
-
-
-
-        "form-title": "Contact me",
-        "success-message": "Thanks for your message! I'll get back to you ASAP.",
-        "name-label": "Name",
-        "email-label": "Email",
-        "message-label": "Message",
-        "submit-button": "Send",
-
-
-        "footerText":{
-                "1": "© 2024 Alejandro Navarro García - Web Developer. All rights reserved.",
-                "2": " Also interested in: Sustainability || ",
-                "3": " Animal Care || ",
-                "4": " Video Game Development || ",
-                "5": " Cinema || ",
-                "6": " Music || "
-        }
-        
-}
