--- public/data/data-es.json
@@ -1,25 +1,0 @@
-{
-        "aboutMe": "Sobre Mi",
-        "studies": "Estudios",
-        "skills": "Skills",
-        "work": "Trabajo",
-        "projects": "Proyectos",
-       "aboutMeTitle": "Hola soy Alex Web developer y diseñador de escenografía!",
-       "aboutMePhrase":"Como recién graduado en Desarrollo Web, me apasiona crear sitios web y aplicaciones centradas en el usuario y con impacto. Durante mis estudios, he desarrollado una sólida base en tecnologías de back-end y front-end. Mi práctica profesional me ha permitido explorar algunas etapas del ciclo de desarrollo, y estoy entusiasmado por aportar mis habilidades y creatividad a un equipo dinámico, además de continuar creciendo como desarrollador.",
-        "studiesTitle": "Estudios",
-        "studies1": {
-                "title": "Estudio 1",
-                "description": "Descripción del estudio 1",
-                "date": "Fecha del estudio 1"
-                },
-        "studies2": {
-                        "title": "Estudio 2",
-                        "description": "Descripción del estudio 1",
-                        "date": "Fecha del estudio 1"
-                },
-        "studies3": {
-                        "title": "Estudio 3",
-                        "description": "Descripción del estudio 1",
-                        "date": "Fecha del estudio 1"
-                }
-}
