+++ public/data/lang-en.json
@@ -1,0 +1,105 @@
+{
+        "aboutMe": "About Me",
+        "studies": "Education",
+        "skills": "Skills",
+        "work": "Experience",
+        "projects": "Projects",
+        "aboutMeTitle": "Hi, I’m Alex!",
+        "aboutMeSubtitle": "Web Developer and Stage Designer",
+        "aboutMePhrase": "As a recent graduate in Web Development, I am passionate about creating impactful, user-focused websites and applications. My expertise spans both front-end and back-end, with a greater focus on Java and MySQL, supported by a solid foundation in design and teamwork principles.",
+        "studiesTitle": "Education",
+        "studies1": {
+            "title": "Higher Technician in Digital Design and Web Development",
+            "description": "Gained proficiency in essential programming languages and successfully developed various projects. My focus was primarily on Java and Spring Boot, along with MySQL for database management and HTML, CSS, and JavaScript for front-end development. Additionally, I worked with version control tools like Git and GitHub, and explored deployment processes using Azure.",
+            "site": "Colegio Santa Joaquina de Vedruna",
+            "date": "2023-2025"
+        },
+        "studies2": {
+            "title": "Bachelor's Degree in Stage Design and Scenography Technology",
+            "description": "Completed a four-year program focused on designing and collaborating on theater projects. Developed expertise in design, production, and team leadership within a creative environment.",
+            "site": "Colegio Santa Joaquina de Vedruna",
+            "date": "2017-2022"
+        },
+        "studies3": {
+            "title": "Intermediate Level French Language Certification",
+            "description": "Achieved a B2 level in French through studies at the Instituto de Idiomas de la Universidad de Sevilla.",
+            "site": "Instituto de Idiomas de la Universidad de Sevilla",
+            "date": "2021"
+        },
+        "studies4": {
+            "title": "Advanced Level English Language Certification",
+            "description": "Attended a bilingual program in Granada and reached a C1 level in English.",
+            "site": "Bilingual School of Granada",
+            "date": "2017"
+        },
+        "skillsTitle": "Skills",
+        "hardskills": "Hard Skills",
+        "softskills": "Soft Skills",
+        "Comunication": "Comunication",
+        "Teamwork": "Teamwork",
+        "Creativity": "Creativity",
+        "Leadership": "Leadership",
+        "Problem Solving": "Problem Solving",
+        "Critical Thinking": "Critical Thinking",
+        "workTitle": "Professional Experience",
+        "work1": {
+            "title": "NTT DATA Internship",
+            "description": "Completed a 3-month dual internship program at NTT DATA as a Junior Web Developer, where I gained hands-on experience in web development and team collaboration.",
+            "date": "2023",
+            "link": "https://es.nttdata.com"
+        },
+        "work2": {
+            "title": "Crew Leader at Veganitessen",
+            "description": "Led a team at Veganitessen, managing daily operations and ensuring excellent service in a fast-paced hospitality environment.",
+            "date": "2020-2021"
+        },
+        "work3": {
+            "title": "Stage Designer at Wallahye Productions",
+            "description": "Contributed to creative stage designs and managed the technical execution of projects as part of a collaborative production team.",
+            "date": "2019-2020"
+        },
+        "referencesTitle": "References",
+        "reference1": {
+            "img": "/references/reference1.png",
+            "title": "Alejandro Téllez",
+            "description": "Training Supervisor at NTT DATA",
+            "contact": "Contact: a.Tellez@nttdata.es"
+        },
+        "reference2": {
+            "img": "/references/reference2.png",
+            "title": "Anabel Reyes",
+            "description": "Coordinator at Veganitessen",
+            "contact": "Contact: anabelReyes@gmail.com"
+        },
+        "reference3": {
+            "img": "/references/reference3.png",
+            "title": "José Navarro",
+            "description": "Team Leader, Wallahye Productions",
+            "contact": "Contact: wallahyeProd@hotmail.es"
+        },
+        "reference4": {
+            "img": "/references/reference4.png",
+            "title": "Joaquín Borrego",
+            "description": "Web Development Degree Tutor at Vedruna",
+            "contact": "Contact: joaquin@telefónica.es"
+        },
+        "projectTitle": "Projects",
+        "buttonNext": "Next",
+        "buttonPrev": "Previous",
+        "buttonSearch": "Search",
+        "form-title": "Contact Me",
+        "success-message": "Thank you for your message! I’ll get back to you as soon as possible.",
+        "name-label": "Name",
+        "email-label": "Email",
+        "message-label": "Message",
+        "submit-button": "Send",
+        "footerText": {
+            "1": "© 2024 Alejandro Navarro García - Web Developer. All rights reserved.",
+            "2": " Also interested in: Sustainability || ",
+            "3": " Animal Care || ",
+            "4": " Video Game Development || ",
+            "5": " Cinema || ",
+            "6": " Music"
+        }
+    }
+
