+++ public/data/lang-es.json
@@ -1,0 +1,107 @@
+{
+        "aboutMe": "Sobre mí",
+        "studies": "Estudios",
+        "skills": "Habilidades",
+        "work": "Experiencia",
+        "projects": "Proyectos",
+        "aboutMeTitle": "¡Hola, soy Alex!",
+        "aboutMeSubtitle": "Desarrollador web y Escenógrafo",
+        "aboutMePhrase": "Como recién graduado en Desarrollo Web, me apasiona crear sitios web y aplicaciones innovadoras, centradas en el usuario. Mi experiencia abarca tanto tecnologías de front-end como de back-end, con un foco en Java y MySQL, respaldada por una sólida base en principios de diseño y trabajo en equipo.",
+        "studiesTitle": "Estudios",
+        "studies1": {
+            "title": "Técnico Superior en Diseño y Desarrollo Web",
+            "description": "Desarrollé competencias en lenguajes de programación esenciales y llevé a cabo varios proyectos. Me especialicé principalmente en Java y Spring Boot, con MySQL para gestión de bases de datos, y HTML, CSS y JavaScript para el desarrollo front-end. También trabajé con herramientas de control de versiones como Git y GitHub, y exploré procesos de implementación usando Azure.",
+            "site": "Colegio Santa Joaquina de Vedruna",
+            "date": "2023-2025"
+        },
+        "studies2": {
+            "title": "Grado en Diseño Escenográfico",
+            "description": "Completé una carrera profesional enfocada en el diseño escenográfico. Desarrollé habilidades en escenografía tecnica, producción y gestión de equipos dentro de un entorno creativo.",
+            "site": "Escuela Superior de Arte Dramárico de Sevilla",
+            "date": "2017-2022"
+        },
+        "studies3": {
+            "title": "Certificación de nivel intermedio en francés",
+            "description": "Alcancé un nivel B2 en francés tras estudiar en el Instituto de Idiomas de la Universidad de Sevilla.",
+            "site": "Instituto de Idiomas de la Universidad de Sevilla",
+            "date": "2021"
+        },
+        "studies4": {
+            "title": "Certificación de nivel avanzado en inglés",
+            "description": "Asistí a un programa bilingüe en Granada y alcancé un nivel C1 en inglés.",
+            "site": "Colegio Bilingüe de Granada",
+            "date": "2017"
+        },
+        "skillsTitle": "Habilidades",
+        "hardskills": "Habilidades Técnicas",
+        "softskills": "Habilidades Interpersonales",
+        "Comunication": "Comunicación",
+        "Teamwork": "Trabajo en Equipo",
+        "Creativity": "Creatividad",
+        "Leadership": "Liderazgo",
+        "Problem Solving": "Resolución de Problemas",
+        "Critical Thinking": "Pensamiento Critico",
+        "workTitle": "Experiencia Profesional",
+        "work1": {
+            "title": "Prácticas en NTT DATA",
+            "description": "Completé un programa de prácticas duales de 3 meses en NTT DATA como Desarrollador Web Junior, adquiriendo experiencia práctica en desarrollo web y trabajo en equipo.",
+            "date": "2023",
+            "link": "https://es.nttdata.com"
+        },
+        "work2": {
+            "title": "Líder de equipo en Veganitessen",
+            "description": "Lideré un equipo en Veganitessen, gestionando horarios  y garantizando un servicio excelente en un entorno universitario dinámico.",
+            "date": "2020-2021",
+            "link":"https://veganitessen.es"
+        },
+        "work3": {
+            "title": "Diseñador Escénico en Wallahye Productions",
+            "description": "Contribuí al diseño escénicos creativo y gestioné la ejecución técnica de proyectos como parte de un equipo de producción colaborativo.",
+            "date": "2019-2020",
+            "link":"https://es.linkedin.com/in/wallahye-productions-sevilla-5ba302163"
+        },
+        "referencesTitle": "Referencias",
+        "reference1": {
+            "img": "/references/reference1.png",
+            "title": "Alejandro Téllez",
+            "description": "Supervisor de formación en NTT DATA",
+            "contact": "Contacto: a.Tellez@nttdata.es"
+        },
+        "reference2": {
+            "img": "/references/reference2.png",
+            "title": "Anabel Reyes",
+            "description": "Coordinadora en Veganitessen",
+            "contact": "Contacto: anabelReyes@gmail.com"
+        },
+        "reference3": {
+            "img": "/references/reference3.png",
+            "title": "José Navarro",
+            "description": "Líder de equipo, Wallahye Productions",
+            "contact": "Contacto: wallahyeProd@hotmail.es"
+        },
+        "reference4": {
+            "img": "/references/reference4.png",
+            "title": "Joaquín Borrego",
+            "description": "Tutor del grado de Desarrollo de Aplicaciones Web en Vedruna",
+            "contact": "Contacto: joaquin@telefónica.es"
+        },
+        "projectTitle": "Proyectos",
+        "buttonNext": "Siguiente",
+        "buttonPrev": "Anterior",
+        "buttonSearch": "Buscar",
+        "form-title": "Contáctame",
+        "success-message": "¡Gracias por tu mensaje! Me pondré en contacto contigo lo antes posible.",
+        "name-label": "Nombre",
+        "email-label": "Correo electrónico",
+        "message-label": "Mensaje",
+        "submit-button": "Enviar",
+        "footerText": {
+            "1": "© 2024 Alejandro Navarro García - Desarrollador Web. Todos los derechos reservados.",
+            "2": " También interesado en: Sostenibilidad || ",
+            "3": " Cuidado Animal || ",
+            "4": " Desarrollo de Videojuegos || ",
+            "5": " Cine || ",
+            "6": " Música"
+        }
+    }
+
