+++ src/components/References.astro
@@ -15,7 +15,7 @@ const { id } = Astro.props;
   />
   <div class="flex flex-col justify-center ml-6 text-whiteBrkn w-2/3">
     <h4 class="text-xl font-montserrat font-bold text-whiteBrkn mb-2" id={`${id}-title`}></h4>
-    <p class="font-montserrat text-sm text-whiteBrkn mb-3" id={`${id}-description`}></p>
+    <p class="font-montserrat text-sm text-whiteBrkn mb-3 text-left" id={`${id}-description`}></p>
     <p class="font-montserrat text-xs text-crudo" id={`${id}-contact`}></p>
   </div>
 </div>
