+++ src/components/studies.astro
@@ -7,13 +7,14 @@ interface props {
 const { id } = Astro.props;
 ---
 
-<div class="flex flex-row py-10 gap-32 border-l-2 border-azulGris border-opacity-70 p-10  transition-all hover:translate-x-2 hover:translate-y-2 hover:shadow-lg hover:bg-paleBlue hover:bg-opacity-40 duration-500 ease-in-out">
-  <div id={id} class="flex justify-left align-middle flex-wrap flex-row w-1/2">
-    <h3 class="text-3xl font-montserrat italic font-semibold text-azulGris text-wrap" id={`${id}-title`}></h3>
-    <p class="font-roboto font-bold italic text-20 text-crudo w-10/12" id={`${id}-date`}></p>
+<div class="flex flex-row py-10 gap-32 border-l-2 border-azulGris border-opacity-70 p-10 justify-center transition-all hover:translate-x-2 hover:translate-y-2 hover:shadow-lg hover:bg-paleBlue hover:bg-opacity-40 duration-500 ease-in-out">
+  <div id={id} class="flex justify-left align-middle flex-wrap flex-row w-1/3 gap-3">
+    <h3 class="text-3xl font-montserrat  italic font-semibold text-darkBlue text-left" id={`${id}-title`}></h3>
+    <p class="font-roboto font-bold italic text-20 text-crudo w-10/12" id={`${id}-site`}></p>
+    <p class="font-roboto font-bold italic text-20 text-darkBlue w-10/12" id={`${id}-date`}></p>
   </div>
-  <div class="flex justify-left align-middle flex-wrap flex-row">
-    <p class="font-montserrat font-medium text-20 text-azulGris w-10/12" id={`${id}-description`}></p>
+  <div class="flex justify-left align-middle flex-wrap flex-row w-1/2">
+    <p class="font-montserrat font-medium text-20 text-darkBlue w-10/12" id={`${id}-description`}></p>
   </div>
 </div>
