+++ src/layouts/LayoutAdmin.astro
@@ -1,3 +1,5 @@
+
+<!-- Layout para las secciones de administración sin elementos como el footer -->
 <html lang="en">
 	<head>
 		<meta charset="utf-8" />
